feature/error_codes Return ESP error codes from slave to app on host
- errors encountered by slave while executing commands are now
  correctly passed to the caller on the host
- clean-up and re-implemented some RPC calls on slave to use macros:
  - get/set mac address: originally mac address is a byte
    string ("aa:bb:cc:dd:ee"). Now data is an array of bytes as
    required by ESP-IDF API
  - get/set wifi mode
  - get/set ps
  - get/set max tx power
- added RCP proto structs for get/set ps
  - was using RPC struct for get/set mode
- added rcp wrapper and esp hosted for get/set max tx power
  - to be called by wifi remote component
- clean up LOG messages
- clean up code
  - removed BSSID_LENGTH
  - used MACSTR and MAC2STR macros
  - removed some unused code

+++ host/drivers/rpc/core/rpc_core.h
@@ -126,7 +126,7 @@ ctrl_cmd_t * rpc_wait_and_parse_sync_resp(ctrl_cmd_t *req);
  * > MSG_ID_OUT_OF_ORDER - if request msg id is unsupported
  * > CALLBACK_NOT_REGISTERED - if aync callback is not available
  **/
-int compose_rpc_req(Rpc *req, ctrl_cmd_t *app_req, uint8_t *failure_status);
+int compose_rpc_req(Rpc *req, ctrl_cmd_t *app_req, int32_t *failure_status);
 
 int is_event_callback_registered(int event);
